v1.6

+++ src/me/taylorkelly/mywarp/Converter.java
@@ -10,6 +10,7 @@ import java.util.Scanner;
 import java.util.logging.Level;
 
 import org.bukkit.*;
+import org.bukkit.entity.Player;
 
 public class Converter {
 
@@ -48,9 +49,9 @@ public class Converter {
 				ps.setString(2, warp.name);
 				ps.setString(3, warp.creator);
 				ps.setInt(4, warp.world);
-				ps.setInt(5, warp.x);
+				ps.setDouble(5, warp.x);
 				ps.setInt(6, warp.y);
-				ps.setInt(7, warp.z);
+				ps.setDouble(7, warp.z);
 				ps.setInt(8, warp.yaw);
 				ps.setInt(9, warp.pitch);
 				ps.setBoolean(10, warp.publicAll);
